Rename: CryptMethod to Cypher
