+++ Forms/MainForm.cs
@@ -8,7 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.IO;
-using DataEncrypter.CryptMethods;
+using DataEncrypter.Cyphers;
 using DataEncrypter.IO;
 
 namespace DataEncrypter
