+++ Readme.md
@@ -1,5 +1,5 @@
 # DataEncrypter
-*DataEncrypter* is an easy to use tool for encryption and decryption. It is build to be lightweight and not memory intensive. It currently has a custom implementation of the Advanced Encryption Standard, but is open to be used with different algorithms as well, if they implement the ICryptMethod interface.
+*DataEncrypter* is an easy to use tool for encryption and decryption. It is build to be lightweight and not memory intensive. It currently has a custom implementation of the Advanced Encryption Standard, but is open to be used with different algorithms as well, if they implement the ICypher interface.
 
 File encryption by using the SecureFile class is very simple to handle. Files are saved as .secf (SECureFile) and provide a encrypted header with previous filename and extension, to restore the file completely after decryption.
